+++ frameworks/react/src/components/select/examples/basic.tsx
@@ -18,8 +18,8 @@ export const Basic = () => {
       <Portal>
         <Select.Positioner>
           <Select.Content>
-            <Select.ItemGroup id="framework">
-              <Select.ItemGroupLabel htmlFor="framework">Frameworks</Select.ItemGroupLabel>
+            <Select.ItemGroup>
+              <Select.ItemGroupLabel>Frameworks</Select.ItemGroupLabel>
               {items.map((item) => (
                 <Select.Item key={item} item={item}>
                   <Select.ItemText>{item}</Select.ItemText>