+++ frameworks/react/src/components/select/select-item-group.tsx
@@ -1,19 +1,26 @@
 import { mergeProps } from '@zag-js/react'
 import type { ItemGroupProps } from '@zag-js/select'
-import { forwardRef } from 'react'
-import type { Assign } from '../../types'
+import { forwardRef, useId } from 'react'
 import { createSplitProps } from '../../utils/create-split-props'
 import { type HTMLArkProps, ark } from '../factory'
 import { useSelectContext } from './use-select-context'
+import { SelectItemGroupPropsProvider } from './use-select-item-group-props'
 
-export interface SelectItemGroupProps extends Assign<HTMLArkProps<'div'>, ItemGroupProps> {}
+export interface SelectItemGroupProps extends HTMLArkProps<'div'> {}
 
 export const SelectItemGroup = forwardRef<HTMLDivElement, SelectItemGroupProps>((props, ref) => {
-  const [itemGroupProps, localProps] = createSplitProps<ItemGroupProps>()(props, ['id'])
+  const id = useId()
+  const [_itemGroupProps, localProps] = createSplitProps<Partial<ItemGroupProps>>()(props, ['id'])
+  const itemGroupProps = { id, ..._itemGroupProps }
+
   const select = useSelectContext()
   const mergedProps = mergeProps(select.getItemGroupProps(itemGroupProps), localProps)
 
-  return <ark.div {...mergedProps} ref={ref} />
+  return (
+    <SelectItemGroupPropsProvider value={itemGroupProps}>
+      <ark.div {...mergedProps} ref={ref} />
+    </SelectItemGroupPropsProvider>
+  )
 })
 
 SelectItemGroup.displayName = 'SelectItemGroup'