+++ frameworks/react/src/components/select/tests/basic.tsx
@@ -28,8 +28,8 @@ export const ComponentUnderTest = (props: Optional<SelectRootProps<Item>, 'items
       <Portal>
         <Select.Positioner data-testid="positioner">
           <Select.Content>
-            <Select.ItemGroup id="framework">
-              <Select.ItemGroupLabel htmlFor="framework">Frameworks</Select.ItemGroupLabel>
+            <Select.ItemGroup>
+              <Select.ItemGroupLabel>Frameworks</Select.ItemGroupLabel>
               {items.map((item) => (
                 <Select.Item key={item.value} item={item}>
                   <Select.ItemText>{item.label}</Select.ItemText>