+++ frameworks/solid/src/components/select/examples/advanced.tsx
@@ -26,8 +26,8 @@ export const Advanced = () => {
       <Portal>
         <Select.Positioner>
           <Select.Content>
-            <Select.ItemGroup id="framework">
-              <Select.ItemGroupLabel for="framework">Frameworks</Select.ItemGroupLabel>
+            <Select.ItemGroup>
+              <Select.ItemGroupLabel>Frameworks</Select.ItemGroupLabel>
               <Index each={items}>
                 {(item) => (
                   <Select.Item item={item()}>