+++ frameworks/solid/src/components/select/examples/basic.tsx
@@ -16,8 +16,8 @@ export const Basic = () => {
       <Portal>
         <Select.Positioner>
           <Select.Content>
-            <Select.ItemGroup id="framework">
-              <Select.ItemGroupLabel for="framework">Frameworks</Select.ItemGroupLabel>
+            <Select.ItemGroup>
+              <Select.ItemGroupLabel>Frameworks</Select.ItemGroupLabel>
               <Index each={items}>
                 {(item) => (
                   <Select.Item item={item()}>