+++ frameworks/solid/src/components/select/select-item-group.tsx
@@ -1,16 +1,22 @@
 import type { ItemGroupProps } from '@zag-js/select'
 import { mergeProps } from '@zag-js/solid'
-import type { Assign } from '../../types'
+import { createUniqueId } from 'solid-js'
 import { createSplitProps } from '../../utils/create-split-props'
 import { type HTMLArkProps, ark } from '../factory'
 import { useSelectContext } from './use-select-context'
+import { SelectItemGroupPropsProvider } from './use-select-item-group-props-context'
 
-export interface SelectItemGroupProps extends Assign<HTMLArkProps<'div'>, ItemGroupProps> {}
+export interface SelectItemGroupProps extends HTMLArkProps<'div'> {}
 
 export const SelectItemGroup = (props: SelectItemGroupProps) => {
-  const [itemGroupProps, localProps] = createSplitProps<ItemGroupProps>()(props, ['id'])
+  const [_itemGroupProps, localProps] = createSplitProps<Partial<ItemGroupProps>>()(props, ['id'])
   const select = useSelectContext()
+  const itemGroupProps = mergeProps({ id: createUniqueId() }, _itemGroupProps)
   const mergedProps = mergeProps(() => select().getItemGroupProps(itemGroupProps), localProps)
 
-  return <ark.div {...mergedProps} />
+  return (
+    <SelectItemGroupPropsProvider value={itemGroupProps}>
+      <ark.div {...mergedProps} />
+    </SelectItemGroupPropsProvider>
+  )
 }