+++ frameworks/vue/src/components/select/examples/basic.vue
@@ -21,8 +21,8 @@ const items = ref(['React', 'Solid', 'Vue'])
     <Teleport to="body">
       <Select.Positioner>
         <Select.Content>
-          <Select.ItemGroup id="framework">
-            <Select.ItemGroupLabel htmlFor="framework">Frameworks</Select.ItemGroupLabel>
+          <Select.ItemGroup>
+            <Select.ItemGroupLabel>Frameworks</Select.ItemGroupLabel>
             <Select.Item v-for="item in items" :key="item" :item="item">
               <Select.ItemText>{{ item }}</Select.ItemText>
               <Select.ItemIndicator>✓</Select.ItemIndicator>