+++ frameworks/vue/src/components/select/tests/select.test.vue
@@ -22,8 +22,8 @@ const items = [
     <Teleport to="body">
       <Select.Positioner data-testid="positioner">
         <Select.Content>
-          <Select.ItemGroup id="framework">
-            <Select.ItemGroupLabel htmlFor="framework">Frameworks</Select.ItemGroupLabel>
+          <Select.ItemGroup>
+            <Select.ItemGroupLabel>Frameworks</Select.ItemGroupLabel>
             <Select.Item v-for="item in items" :key="item.value" :item="item">
               <Select.ItemText>{{ item.label }}</Select.ItemText>
               <Select.ItemIndicator>✓</Select.ItemIndicator>