+++ frameworks/vue/src/components/select/use-select-item-group-props-context.ts
@@ -1,0 +1,7 @@
+import type { ItemGroupProps } from '@zag-js/select'
+import { createContext } from '../../utils'
+
+export interface UseSelectItemGroupPropsContext extends ItemGroupProps {}
+
+export const [SelectItemGroupPropsProvider, useSelectItemGroupPropsContext] =
+  createContext<UseSelectItemGroupPropsContext>('SelectItemGroupPropsContext')